mocks saves to user output dir; moved walk.go into filewalk package; moved hash comparasion login into Run func
